add accounts screen
Step 1 of a few to implement full account functionality through
the UI. Refactors a lot of UI code and simplifies the Wallet
API. Also adds a SimpleWallet type, which is a single-account Decred
wallet.

Adds an account selection screen. Right now, there is still only
one account available, and no option to create more. That will
follow in the next step.