add dcrd-based Blockchain class
Get the basic structure in place for a class called LocalNode that
is intended to be interchangable with DcrdataBlockchain, though far
from complete. It uses the new rpc.WebsocketClient to access the
websocket-only endpoints of the dcrd RPC. A few basic functions
used for syncing accounts are in place to demonstrate how
AccountManager and Account would use LocalNode.