tinywallet: connect to dcrd JSON-RPC server
Adds screen and settings necessary to connect to dcrd. The screen
displays either a form to enter server credential, of if already
configured, a screen to turn the connection on or off or to
forget the connection.

When the LocalNode is connected or disconnected, the Accounts are
informed. They store the current LocalNode, but don't actually use
it yet.