app: Add relay fee setting
Add a relay fee to settings and allow changing. The fee is not yet used.

Save the relay fee with the account. Allow changing of the relay fee in
Account settings as apposed to App settings. Change relay fee input
field to radio buttons with labels low, default, and high. Set
appropriate values for low (4000) and high (114000).