screens: Continue on open file cancel. (#183)
Rather than throwing an exception when restoring a wallet from file,
just return, allowing the user to do something else.

Create a logging helper for errors that logs to stdErr and the main ui
window.

Do not move a user's wallet file without warning. Copy it instead.

Check that the file is a sqlite database, otherwise you get stuck in a
loop until you delete or move the file manually.